+++ Qlik.Sense.RestClient/Qlik.Sense.RestClient/RestClientDebugConsole.cs
@@ -37,17 +37,8 @@ namespace Qlik.Sense.RestClient
         {
             if (disposing)
             {
-                RestClient.RestClientDebugConsole = null;
+                RestClientGeneric.RestClientDebugConsole = null;
             }
-        }
-    }
-
-    [Obsolete("Class deprecated due to name clash with Qlik Sense .NET SDK. Use class RestClientDebugConsole instead.")] // Obsolete since January 2022
-    public class DebugConsole : RestClientDebugConsole
-    {
-        [Obsolete("Class deprecated due to name clash with Qlik Sense .NET SDK. Use class RestClientDebugConsole instead.")] // Obsolete since January 2022
-        public DebugConsole() : base()
-        {
         }
     }
 }